requirements.txt
pandas
yfinance
ipython
ipythonblocks
ipywidgets
jupyter
numpy
discord.py
python-dotenv
yahoofinancials
schedule